[fix] 提现逻辑修改

+++ src/main/java/com/tianli/chain/dto/BaseTokenReq.java
@@ -3,6 +3,7 @@ package com.tianli.chain.dto;
 import lombok.Data;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
@@ -24,7 +25,7 @@ public class BaseTokenReq {
 
     private String block;
 
-    private BigDecimal value;
+    private BigInteger value;
 
     private String contractAddress;
